Follow-up of `moc.js` API update (Serokell, Milestone-1)  (#113)
* [VSC-7] Change moc.js functions for the LSP

Problem: In the compiler, we changed the `parseMotokoTyped` function for
the LSP that allows caching for better performance. This needs to be
exposed by node-motoko. Moreover, `parseMotokoWithDeps` was added which
parses the file and returns immediate dependencies.

Solution: Regenerate based on a local compiler build. Change `file.ts`
and `index.ts` to account for these functions.

* [VSC-7] Add a couple of tests

Problem: We should test our changes related to caching, to ensure
everything is working as intended.

Solution: Write a couple of tests related to checking caches for changed
and unchanged files.

* [VSC-7] Fix dependency analysis

Problem: We need to pass the filename to the import resolver, which
requires passing the filename to `parseMotokoWithDeps`.

Solution: Regenerate compiler and use the updated compiler API. Add some
tests to be sure it's working.

* [VSC-7] Make scopes cache optional

Problem: We were asked during review to make scope cache optional.

Solution: This is accompanied by some changes in the compiler as well.
Turn `scopeCache: ...` argument into `scopeCache?: ...`.

* [VSC-7] Improve backwards compatibility

Problem: We need to improve backwards compability: if no scope is
passed, the return type should have no scope.

Solution: Create more overloads for `parseMotokoTyped` and write a test
checking that it's working as intended.

* [VSC-7] Explain backwards compatible code

Problem: It's a little confusing why we do the arguments.length check,
we should explain it.

Solution: Add a comment in both overloads explaining this design.

* Revert last three commits

* [VSC-20] Split parseMotokoTyped

* [VSC-20] Add a test for new parseMotokoTyped

Problem: We don't have a test for parseMotokoTyped.

Solution: Add it.

* [VSC-7] Fix hardcoded version in tests

Problem: We check for `moc-0.14.1` in one of our tests, but this will
break if the compiler version changes.

Solution: Match the version using regex instead.

* Update moc to 0.14.7

* update moc to 0.14.7
* bump version to 3.11.7

---------

Co-authored-by: Heitor Toledo Lassarote de Paula <heitor.toledo@serokell.io>
Co-authored-by: Savely Krendelhoff <alicecandyom@proton.me>